fe-streamlit/requirements.txt
streamlit==1.47.1
python-dotenv
requests